preserve ticket basics in Jumbo
so when we are clicking "Add More Files" or "Go"es to search people or groups,
the changes we made could be preserved.

note that I moved the handling "2 Owners" code out of the condition block so
it can be handled always. without this, the owner input can't be successfully
preserved.  (i.e. you could get an empty default value no matter which owner
you chose)