it's not RT::Ticket::Forward's job to encode headers to mime
previously if the subject of forward email contains non-ascii charts, it will
be encoded to mime like "=?UTF-8?B?w6TDtsO8?="(i.e. "äöü") in ->Forward, which
is wrong because we want utf8 data there so in RT::Action::SendEmail, we can
massage it(e.g. to add subject token) and _then_ encode it to mime.

without this, non-ascii subject would be _double_ mime encoded if subject tag
also contains non-ascii charts.

another issue is: non-ascii subject of forwarded emails would be saved as mime
encoded string(like "=?UTF-8?B?w6TDtsO8?=") in Attachments table(in both Subject
and Headers columns) and be shown the _same_ way in history part of ticket
display page.

Fixes: I#29714