Drop the "internal formatter" message from info to debug.
An informational message added in 40ba502 appears rather frequently
in normal RT operation (once per commandline script in some cases)
and isn't relevant to audit logs or directly actionable.

If this is important enough to log about on every start of the
server and every commandline script, we may want to consider making
it a warning and recommending corrective action in the text.